requirements.txt
grpcio==0.15
grpcio-tools==0.15.1
pre-commit==0.8.2
python-gflags==3.0.5
